// src/entities/Weather/models/types/weather.ts
export interface Weather {
    cod: string;
    message: number;
    cnt: number;
    list: WeatherData[];
    city: WeatherCity;
}

export interface WeatherData {
    dt: number;
    main: {
        temp: number;
        feels_like: number;
        temp_min: number;
        temp_max: number;
        pressure: number;
        sea_level: number;
        grnd_level: number;
        humidity: number;
        temp_kf: number;
    };
    weather: WeatherDescription[];
    clouds: {
        all: number;
    };
    wind: {
        speed: number;
        deg: number;
        gust: number;
    };
    visibility: number;
    pop: number;
    sys: {
        pod: string;
    };
    dt_txt: string;
}

interface WeatherDescription {
    id: number;
    main: string;
    description: string;
    icon: string;
}

interface WeatherCity {
    id: number;
    name: string;
    coord: {
        lat: number;
        lon: number;
    };
    country: string;
    population: number;
    timezone: number;
    sunrise: number;
    sunset: number;
}